tests/requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.11
# by the following command:
#
#    pip-compile '.\requirements.in'
#
anyio==4.3.0
    # via httpx
certifi==2024.2.2
    # via
    #   httpcore
    #   httpx
colorama==0.4.6
    # via pytest
h11==0.14.0
    # via httpcore
httpcore==1.0.5
    # via httpx
httpx==0.27.0
    # via -r .\requirements.in
idna==3.7
    # via
    #   anyio
    #   httpx
iniconfig==2.0.0
    # via pytest
packaging==24.0
    # via pytest
pluggy==1.4.0
    # via pytest
pytest==8.1.1
    # via -r .\requirements.in
python-dotenv==1.0.1
    # via -r .\requirements.in
sniffio==1.3.1
    # via
    #   anyio
    #   httpx
